+++ CHANGES.md
@@ -4,6 +4,9 @@
 
 ### misc
 
+- [ADD] notify と push と signalingNotifyMetadata の E2E テストを追加する
+  - @zztkm
+
 - [ADD] devtools を Windows でも使えるようにする
   - @zztkm